pyproject.toml
[tool.poetry]
name = "hose_sound"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.5"
django = "^2.1"
psycopg2 = "^2.7"
jinja2 = "^2.10"
djangorestframework = "^3.9"
django-webpack-loader = "^0.6.0"
django-cors-headers = "^2.4"
djangorestframework_jwt = "^1.11"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
ipython = "^7.2"
pytest = "^4.1"
pytest-django = "^3.4"
pytest-xdist = "^1.26"
pgcli = "^2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
